# phase3-django-api/swagger.yaml
openapi: 3.0.3
info:
  title: Library Management System API
  version: 1.0.0
  description: "\nA RESTful API for managing library operations including:\n\n• Library
    branches  \n• Books, authors, and categories  \n• Member registration and management
    \ \n• Borrowing and returning books  \n• Book reviews and ratings  \n• Library
    analytics and statistics  \n\nThis API supports advanced features such as:\n\n•
    Book search by author, category, and library  \n• Book availability checks  \n•
    Borrowing history tracking  \n• Recommendation engine for popular and highly rated
    books  \n• Aggregated statistics for library usage  \n\nAll endpoints return JSON
    responses and follow RESTful conventions.\n"
paths:
...
  /api/borrowings/:
    get:
      operationId: borrowings_list
      description: Retrieve a paginated list of all borrowing transactions.
      summary: List Borrowings
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - borrowings
      security:
      - cookieAuth: []
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBorrowingList'
              examples:
                BorrowingExample:
                  value:
                    count: 123
                    next: http://api.example.org/accounts/?page=4
                    previous: http://api.example.org/accounts/?page=2
                    results:
                    - borrowing_id: 10
                      member_name: John Doe
                      member:
                        member_id: 5
                        first_name: John
                        last_name: Doe
                      book:
                        book_id: 1
                        title: Harry Potter and the Philosopher's Stone
                        isbn: '9780747532699'
                      borrow_date: '2026-03-10'
                      due_date: '2026-03-24'
                      return_date: null
                      late_fee: null
                  summary: Example of a borrowing transaction
                  description: Represents a book borrowing transaction.
          description: ''
    post:
      operationId: borrowings_create
      description: Create a new borrowing transaction directly.
      summary: Create Borrowing
      tags:
      - borrowings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Borrowing'
            examples:
              BorrowingExample:
                value:
                  borrowing_id: 10
                  member_name: John Doe
                  member:
                    member_id: 5
                    first_name: John
                    last_name: Doe
                  book:
                    book_id: 1
                    title: Harry Potter and the Philosopher's Stone
                    isbn: '9780747532699'
                  borrow_date: '2026-03-10'
                  due_date: '2026-03-24'
                  return_date: null
                  late_fee: null
                summary: Example of a borrowing transaction
                description: Represents a book borrowing transaction.
        required: true
      security:
      - cookieAuth: []
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrowing'
              examples:
                BorrowingExample:
                  value:
                    borrowing_id: 10
                    member_name: John Doe
                    member:
                      member_id: 5
                      first_name: John
                      last_name: Doe
                    book:
                      book_id: 1
                      title: Harry Potter and the Philosopher's Stone
                      isbn: '9780747532699'
                    borrow_date: '2026-03-10'
                    due_date: '2026-03-24'
                    return_date: null
                    late_fee: null
                  summary: Example of a borrowing transaction
                  description: Represents a book borrowing transaction.
          description: ''
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        url:
          type: string
          format: uri
          readOnly: true
        library:
          allOf:
          - $ref: '#/components/schemas/NestedLibrary'
          readOnly: true
        authors:
          type: array
          items:
            $ref: '#/components/schemas/NestedAuthor'
          readOnly: true
        categories:
          type: array
          items:
            $ref: '#/components/schemas/NestedCategory'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the record was created.
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the record was last updated.
        title:
          type: string
          description: Title of the book.
          maxLength: 50
        isbn:
          type: string
          description: International Standard Book Number.
          maxLength: 15
        publication_date:
          type: string
          format: date
          description: Date the book was originally published.
        total_copies:
          type: integer
          maximum: 2147483647
          minimum: 0
          description: Total number of copies owned by the library.
        available_copies:
          type: integer
          maximum: 2147483647
          minimum: 0
          description: Number of copies currently available for borrowing.
        library_id:
          type: string
          format: uri
          description: Library where this book is located.
      required:
      - authors
      - available_copies
      - categories
      - created_at
      - isbn
      - library
      - library_id
      - publication_date
      - title
      - total_copies
      - updated_at
      - url
...
    Borrowing:
      type: object
      properties:
        url:
          type: string
          format: uri
          readOnly: true
        book:
          allOf:
          - $ref: '#/components/schemas/Book'
          readOnly: true
        member:
          allOf:
          - $ref: '#/components/schemas/NestedMember'
          readOnly: true
        member_name:
          type: string
          readOnly: true
          description: Full name of the member who borrowed the book.
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the record was created.
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the record was last updated.
        borrow_date:
          type: string
          format: date
          description: Date when the book was borrowed.
        due_date:
          type: string
          format: date
          description: Date by which the book must be returned.
        return_date:
          type: string
          format: date
          nullable: true
          description: Actual return date of the book.
        late_fee:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
          description: Late fee charged if returned after due date.
        member_id:
          type: string
          format: uri
          description: Member who borrowed the book.
        book_id:
          type: string
          format: uri
          description: Book that was borrowed.
      required:
      - book
      - book_id
      - created_at
      - due_date
      - member
      - member_id
      - member_name
      - updated_at
      - url
...
    NestedAuthor:
      type: object
      properties:
        author_id:
          type: integer
          readOnly: true
          description: Unique identifier for the author.
        first_name:
          type: string
          description: Author's first name.
          maxLength: 20
        last_name:
          type: string
          description: Author's last name.
          maxLength: 20
      required:
      - author_id
      - first_name
      - last_name
    NestedCategory:
      type: object
      properties:
        category_id:
          type: integer
          readOnly: true
          description: Unique identifier for the category.
        name:
          type: string
          description: Category or genre name.
          maxLength: 30
      required:
      - category_id
      - name
    NestedLibrary:
      type: object
      properties:
        library_id:
          type: integer
          readOnly: true
          description: Unique identifier for the library.
        name:
          type: string
          description: Name of the library.
          maxLength: 30
        campus_location:
          type: string
          description: Campus location where the library is situated.
          maxLength: 30
      required:
      - campus_location
      - library_id
      - name
    NestedMember:
      type: object
      properties:
        member_id:
          type: integer
          readOnly: true
          description: Unique identifier for the member.
        first_name:
          type: string
          description: Member's first name.
          maxLength: 20
        last_name:
          type: string
          description: Member's last name.
          maxLength: 20
      required:
      - first_name
      - last_name
      - member_id
...
    PaginatedBorrowingList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Borrowing'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...